pyproject.toml
[tool.poetry]
name = "eciespy"
packages = [
  {include = "ecies"},
]
version = "0.4.6"
# docs
authors = ["Weiliang Li <to.be.impressive@gmail.com>"]
description = "Elliptic Curve Integrated Encryption Scheme for secp256k1/curve25519 in Python"
license = "MIT"
maintainers = ["Weiliang Li <to.be.impressive@gmail.com>"]
readme = "README.md"
repository = "https://github.com/ecies/py"
# tags
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Natural Language :: English",
  "Programming Language :: Python :: Implementation :: CPython",
  "Operating System :: OS Independent",
  "Topic :: Security :: Cryptography",
]
keywords = [
  "secp256k1",
  "crypto",
  "elliptic curves",
  "ecies",
  "bitcoin",
  "ethereum",
  "cryptocurrency",
]
# package data
include = ["ecies/py.typed"]

[tool.poetry.dependencies]
python = "^3.9"

# for @deprecated
typing-extensions = ">=4.9.0"

# 3rd party
coincurve = ">=13,<22"
pycryptodome = ">=3.21.0"

# optional
eth-keys = {version = ">=0.4,<0.8", optional = true}

[tool.poetry.extras]
eth = ["eth-keys"]

[tool.poetry.group.dev.dependencies]
ipython = {version = "^9.4.0", python = "^3.11"}
mypy = {version = "^1.16.0", python = "^3.11"}
ruff = "^0.12.2"

eth-typing = "^5.2.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.4.0"
pytest-cov = "^6.2.1"

[tool.poetry.scripts]
eciespy = "ecies.__main__:main"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.pytest.ini_options]
addopts = "--doctest-modules --cov=ecies --cov-report term:skip-covered"

[tool.coverage.run]
omit = ["ecies/__main__.py"]

[tool.coverage.report]
exclude_also = ["if TYPE_CHECKING:", "raise NotImplementedError"]
